Added a bunch of improvements to the flutter app (#50)
* Fixed an issue with the watch button

* Initial implementation of caching

* Cleanup references

* More cleanup

* Added caching to the user watchlist

cache gets removed when user changes it... duh makes sense

* Password reset implemented

* Added tutorial page

* Added the disclaimer page

* Added navigation from disclaimer and tutorial

+++ Apps/gm-flutter-app/gimmillions/lib/models/stock-data.dart
@@ -21,6 +21,6 @@ class StockData {
 
   factory StockData.fromJson(Map<String, dynamic> json) {
     return StockData(DateTime.parse(json['date']), json['symbol'], json['open'], json['high'], json['low'],
-        json['close'], json['adjustedClose'], json['volume'], json['previousClose']);
+        json['close'], json['adjustedClose'], json['volume'] as double, json['previousClose']);
   }
 }
